requirements.txt
python-dotenv
ipykernel
pandas
langchain
langchain-openai
rich
streamlit